requirements.txt
# Core dependencies for DeepSeek Validation Tool

# HTTP requests and API interactions
requests>=2.28.0

# Environment variable management
python-dotenv>=1.0.0

# Flask web framework
Flask>=2.3.0

# Email integration (optional)
# Uncomment if using Resend email alerts
# resend>=0.6.0